fix(lux/core): More efficient datastore, dirty fixes
Datastore
Flips the storing mechanism around: From id-store with domain-field journal to
domain-field store with id-journal. This is more efficient when querying, as
excessive looping for every single query is not needed anymore.

This introduced a multiple dirty bugs made by the folks over at luxtronik,
which had to mitigated.

Bloat
Even when requesting a specific set of data, the luxtronik seems to throw more
requested on the client. We don't know what it sends, as it does not mark it, it
only sends unknown ids with usually zero values. Whatever. Ignored those.

Double Usage
The luxtronik seems to use names multiple times. This is a problem, as they must
be unique for the new data store system, otherwise values might get lost.
Fixed this by introducing pluggable filter lambdas, which get all known data and
decide whether to modify the domain or field. This is especially useful for
removing the units later on.